third_party/blink/web_tests/wpt_internal/prerender/windowclient-navigate.https.html

<!--
  This file cannot be upstreamed to WPT until:
  * startPrerendering() usage is replaced with a WebDriver API.
  * The specification changes to disallow navigations after the initial
  prerender navigation.
-->
<!DOCTYPE html>
<title>WindowClient.navigate() for prerendered main page</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<body>
<script>
const PAGE_URL = 'resources/prerendered-page.html?prerendering';
const WORKER_URL = 'resources/windowclient-navigate-worker.js';

promise_test(async t => {
  const registration =
    await service_worker_unregister_and_register(t, WORKER_URL, PAGE_URL);
  t.add_cleanup(() => registration.unregister());
  await wait_for_state(t, get_newest_worker(registration), 'activated');

  const bc = new BroadcastChannel('prerender-channel');
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  startPrerendering(PAGE_URL);

  const prerenderResult = await gotMessage;
  assert_equals(prerenderResult, 'prerender success',
                'should succeed to prerender a page');

  const navigationResult = await new Promise(resolve => {
    navigator.serviceWorker.onmessage = e => { resolve(e.data) };
    registration.active.postMessage('empty.html');
  });
  assert_equals(navigationResult, 'navigate() failed with TypeError',
                'should fail the navigation with TypeError');

}, 'WindowClient.navigate() for a prerendered main page should throw a' +
   'TypeError');
</script>
</body>
